Added logic for last question to properly display the appropriate button in the UX:
if (((totalCorrect + totalIncorrect)==totalQuestions) && (totalQuestions>0))

This insures that the initialization/reset conditions don't interfere with the page refresh or initial load when 0+0=0 for the variables involved.

Next need to add error checking for blank answers, or disallow a response that includes anything but numerical (integer?) values.

Increase font size in question div, make the numbers and text entry field bigger

Begin to create logic for adding division, adding a way for user to set:

-how many questions
-range of values being multiplied
-timed test
-fractions
-graphical percentages (a circle that fills 3/4, How much is filled/empty/given color/etc.)
-database with best score for users
-user login
-update rejoinders to use user name
-add animation for equation to appear to be a flashcard and on answer submit, card flips over
-add variables to track how many tries it took user to get correct and add to results output